Fix kernel version selection using build machine rather than target
This fix uses the 'setup.cfg' file used by setuptools. It reads an
Adafruit-specific section that is ignored by setuptools.

The reason for using the config file is that command-line parameters can't
be passed to setup.py, as the command-line is ideally opaque to the
installer. 'setup.cfg' was added to setuptools as a useful middle-ground
between modifying 'setup.py' and the command-line. Now, library build
systems can modify a config file instead of the source directly.

Fixes https://github.com/adafruit/adafruit-beaglebone-io-python/issues/256